requirements.txt
tensorflow
numpy
scikit-learn
opencv-python
pillow
tqdm
